# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Whim 单组织 RBAC API",
    "version": "1.0.0",
    "description": "用户全局唯一，单主部门，多角色、多岗位。数据范围：1全部、2自定义部门、3本部门、4本部门及下级、5本人。岗位不会隐式授予角色。Long 超出 JavaScript 安全整数时序列化为字符串。"
  },
  "servers": [
    {
      "url": "http://localhost:8089"
    }
  ],
...
  "paths": {
...
    "/system/user": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "分页查询本次操作可见用户。",
        "operationId": "pageUsers",
        "description": "操作权限：`system:user:list`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "页码"
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "每页数量"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageDataVOUserVO"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "用户"
        ],
        "summary": "创建账号，角色和岗位通过独立操作分配。",
        "operationId": "createUser",
        "description": "操作权限：`system:user:create`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultId"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          }
        }
      }
    },
    "/system/user/{userId}": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "查询本次操作可见用户详情。",
        "operationId": "getUser",
        "description": "操作权限：`system:user:detail`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultUserVO"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "修改用户基础信息。",
        "operationId": "updateUser",
        "description": "操作权限：`system:user:update`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "用户"
        ],
        "summary": "删除用户并清理角色和岗位关联。",
        "operationId": "deleteUser",
        "description": "操作权限：`system:user:delete`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      }
    },
    "/system/user/{userId}/department": {
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "分配用户主部门，重新核验迁移后的角色数据范围。",
        "operationId": "setDepartment",
        "description": "操作权限：`system:user:department`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDepartmentRequest"
              }
            }
          }
        }
      }
    },
    "/system/user/{userId}/status": {
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "修改用户状态并撤销旧会话。",
        "operationId": "setUserStatus",
        "description": "操作权限：`system:user:status`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          }
        }
      }
    },
    "/system/user/{userId}/password": {
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "独立授权重置密码，成功后撤销全部旧会话。",
        "operationId": "resetPassword",
        "description": "操作权限：`system:user:resetPassword`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPasswordRequest"
              }
            }
          }
        }
      }
    },
    "/system/user/{userId}/roles": {
      "get": {
        "tags": [
          "用户"
        ],
        "summary": "查询本次操作可见用户的角色ID。",
        "operationId": "getUserRoleIds",
        "description": "操作权限：`system:userRole:list`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultIdList"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "用户"
        ],
        "summary": "覆盖用户角色并核验每个候选角色的委派范围。",
        "operationId": "replaceUserRoles",
        "description": "操作权限：`system:userRole:assign`。按该操作实际获授权的角色合并数据范围，详情、修改和删除均检查目标范围。",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "description": "目标主键 ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          },
          "400": {
            "description": "参数校验或业务规则不满足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "403": {
            "description": "缺少操作权限或目标数据不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "唯一键冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRoleAssignRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "UserCreateRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 32,
            "pattern": "[A-Za-z0-9][A-Za-z0-9._@-]*"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72
          },
          "deptId": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "name": {
            "type": "string",
            "maxLength": 24,
            "minLength": 1
          },
          "email": {
            "type": "string",
            "maxLength": 64,
            "format": "email"
          },
          "phone": {
            "type": "string",
            "maxLength": 20
          },
          "gender": {
            "type": "integer",
            "format": "int32",
            "minimum": 0,
            "maximum": 2,
            "default": 0
          },
          "remark": {
            "type": "string",
            "maxLength": 255
          }
        },
        "required": [
          "username",
          "password",
          "deptId",
          "name"
        ],
        "description": "创建系统账号及主部门，角色和岗位通过独立授权接口分配。"
      },
...
      "ErrorResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "与 HTTP 状态码一致。"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true,
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            ]
          },
          "success": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "code",
          "message",
          "success"
        ]
      },
...
      "ResultId": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "oneOf": [
              {
                "type": "integer",
                "format": "int64"
              },
              {
                "type": "string",
                "pattern": "^-?[0-9]+$"
              }
            ],
            "description": "Java Long；响应超出 JavaScript 安全整数范围时为字符串，请保持 ID 原始值。"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
...
      "ResultPageDataVOUserVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageUserData"
          },
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "success"
        ]
      },
...
    }
  }
}